Move artist recommendations to overview tab of artist page
